Fixed issue that sometimes caused status block not to get called.
This issue occurred in an app that called SCNetworkReachability.initWithHost:
from application:didFinishLaunchingWithOptions:.
In SCNetworkReachability.initWithReachabilityRef: the method
SCReachabilityScheduler.initWithReachabilityRef: is called before
setting statusChangesBlock with method observeStatusChanges:.
In some cases SCReachabilityScheduler.callbackForReachabilityRef
callback function could get called before observeStatusChanges:
got called -> statusChangesBlock was nil and the callback
got ignored. So the app didn't get notified of the network status
until it changed again. This issue was quite easy to reproduce
with an iPhone4 (7.1.2) with WiFi connection (no SIM card was inserted).

The fix was to pass the statusChangesBlock in the
SCReachabilityScheduler.initWithReachabilityRef method.